context data moved away from context file
