+++ tests/core/test_context_dict.py
@@ -1,6 +1,6 @@
 import pytest
 
-from chatsky.core.context import ContextMainInfo
+from chatsky.core.ctx_utils import ContextMainInfo
 from chatsky.context_storages import MemoryContextStorage
 from chatsky.context_storages.database import NameConfig
 from chatsky.core.message import Message
